gh-2774 Refactoring complete
